+++ frontend/src/components/pages/rooms/HostWaitingRoom.tsx
@@ -1,7 +1,8 @@
 import { PushButton } from "@/components/ui/PushButton";
 import { Card } from "flowbite-react";
 import Image from "next/image";
-import { ParticipantList } from "./ParticipantList";
+import { useParams } from "next/navigation";
+import { ParticipantList } from "../../shared/ParticipantList";
 
 interface Participant {
   id: string;
@@ -10,28 +11,36 @@ interface Participant {
   isGuest: boolean;
 }
 
-interface WaitingRoomProps {
-  roomNumber: string;
-  participants: Participant[];
-  onStartQuiz: () => void;
-  onCopyRoomCode: () => void;
-}
+export function HostWaitingRoom() {
+  const { roomCode } = useParams();
+  // const [participants, setParticipants] = useState<Participant[]>([]);
 
-// TODO username avatar
+  // TODO : Demo URL
+  const roomUrl = "/test";
 
-export function HostWaitingRoom({
-  roomNumber,
-  participants,
-  onStartQuiz,
-  onCopyRoomCode,
-}: WaitingRoomProps) {
+  // TODO : Demo data for participants
+  const demoParticipants: Participant[] = [
+    { id: "1", name: "Alice", avatar: "koala", isGuest: false },
+    { id: "2", name: "Bob", avatar: "owl-1", isGuest: true },
+  ];
+
+  const handleStartQuiz = () => {
+    // if (participants.length < 2) {
+    //   return;
+    // }
+  };
+
+  const onCopyRoomCode = () => {
+    navigator.clipboard.writeText(roomCode as string);
+  };
+
   return (
     <div className="container h-full mx-auto px-4 py-8">
       <Card className="max-w-2xl mx-auto">
         <div className="text-center mb-6">
           <h2 className="text-2xl font-bold mb-2">Ready to Start the Quiz</h2>
           <div className="flex items-center justify-center space-x-2">
-            <p className="text-gray-600">Room URL: {roomNumber}</p>
+            <p className="text-gray-600">Room URL: {roomUrl}</p>
             <button
               onClick={onCopyRoomCode}
               className="text-gray-500 hover:text-gray-700 transition-colors p-1 rounded-full active:bg-gray-100"
@@ -44,15 +53,15 @@ export function HostWaitingRoom({
         <div className="mb-6">
           <h3 className="text-xl font-semibold mb-4 text-center">Participants</h3>
           <div className="bg-gray-50 rounded-lg p-4">
-            <ParticipantList participants={participants} />
+            <ParticipantList participants={demoParticipants} />
           </div>
         </div>
         <div className="flex justify-center">
           <PushButton
             color="primary"
             size="md"
-            onClick={onStartQuiz}
-            disabled={participants.length < 2}
+            onClick={handleStartQuiz}
+            // disabled={participants.length < 2}
           >
             Start Quiz
           </PushButton>
